storage/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
aiofiles>=24.0.0
httpx>=0.28.0
requests>=2.32.0
anthropic>=0.40.0
Pillow>=10.0.0
